+++ CHANGES.md
@@ -1,4 +1,5 @@
 #5.0.0.beta1 20??-??-??
+ - 2015-07-13 Updated CPAN module JavaScript::Minifier to version 1.14.
  - 2015-07-13 Updated CPAN module Excel::Writer::XLSX to version 0.84.
  - 2015-07-13 Updated CPAN module Encode::Locale to version 1.05.
  - 2015-07-13 Updated CPAN module Email::Valid to version 1.196.